requirements.txt
fastapi
uvicorn
qwhale_client
python-dotenv
redis